Add create recruiter feature
adds code to routes/sign-up.js and app/services/profiles.js.

When signing up, if the user selects recruiter, a recruiter will
be created in the database for them. Using an if statement in the
promise chain in sing-up.js

next steps: transition to the right route after creating an engineer
or recruiter in the database.