Adds components display conversations for engineer
Ran ember g component create lunch-invite and
lunch-invite/invite

Add a model hook for routes/engineer.js.
The recruiters name and website are displayed for each conversation.
Accessed with conversation.engineer.web and conversation.engineer.name.

Next steps: generate a route to open up the conversation and a button
to link to it